Revert "Merge pull request #91 from sjin11125/cosmo"
This reverts commit d356a1d7dbf1da2e6f9867789456f207e603ba77, reversing
changes made to a5428f7d817113aee591d63d9fe46358c3b34cf0.
