+++ Assets/Scenes/Game.unity
@@ -4406,6 +4406,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 77d974af08f04014d8317a1eafb7afe1, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  QuestActive: 010000000000000000000000
   number: {fileID: 1947962210}
 --- !u!1 &856182679
 GameObject:
@@ -4639,6 +4640,7 @@ RectTransform:
   - {fileID: 300930098}
   - {fileID: 2136930552789152843}
   - {fileID: 2029546643}
+  - {fileID: 3629368370832220583}
   m_Father: {fileID: 0}
   m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -12393,7 +12395,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: 4663d28309109446788af134ecaf938e, type: 3}
+  m_Sprite: {fileID: 21300000, guid: fa4f61c3330b447e2a76c59164fda27e, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -12524,7 +12526,7 @@ MonoBehaviour:
   m_OnCullStateChanged:
     m_PersistentCalls:
       m_Calls: []
-  m_Sprite: {fileID: 21300000, guid: 16f91cf3fb2cb482d847afe440ca0d55, type: 3}
+  m_Sprite: {fileID: 21300000, guid: 5b36194aa5e14456f9ccf3323f15a32d, type: 3}
   m_Type: 0
   m_PreserveAspect: 0
   m_FillCenter: 1
@@ -12903,7 +12905,9 @@ MonoBehaviour:
           m_StringArgument: 
           m_BoolArgument: 0
         m_CallState: 2
+
       - m_Target: {fileID: 2136930552789152840}
+
         m_TargetAssemblyTypeName: UnityEngine.GameObject, UnityEngine
         m_MethodName: SetActive
         m_Mode: 6
@@ -14661,6 +14665,16 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 5920199019524207047}
   m_CullTransparentMesh: 1
+--- !u!1 &3629368370832220582 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 5225903807619729196, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+  m_PrefabInstance: {fileID: 8851887322710877834}
+  m_PrefabAsset: {fileID: 0}
+--- !u!224 &3629368370832220583 stripped
+RectTransform:
+  m_CorrespondingSourceObject: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+  m_PrefabInstance: {fileID: 8851887322710877834}
+  m_PrefabAsset: {fileID: 0}
 --- !u!224 &3677370711936314204
 RectTransform:
   m_ObjectHideFlags: 0
@@ -16156,6 +16170,107 @@ MonoBehaviour:
   m_FillOrigin: 0
   m_UseSpriteMesh: 0
   m_PixelsPerUnitMultiplier: 1
+--- !u!1001 &8851887322710877834
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    m_TransformParent: {fileID: 859097266}
+    m_Modifications:
+    - target: {fileID: 5225903807619729196, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_Name
+      value: infoPanel2
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729196, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_IsActive
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_Pivot.x
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_Pivot.y
+      value: 0.5
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_RootOrder
+      value: 11
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchorMax.x
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchorMax.y
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchorMin.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchorMin.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_SizeDelta.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_SizeDelta.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchoredPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_AnchoredPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 5225903807619729197, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: ccb37e19c2d76174195443c83b7600e7, type: 3}
 --- !u!224 &9056637407577620107
 RectTransform:
   m_ObjectHideFlags: 0
