+++ Assets/Scripts/Game/CreateKeepShape.cs
@@ -75,7 +75,6 @@ public class CreateKeepShape : MonoBehaviour, IPointerDownHandler,IBeginDragHand
     }
     public void OnPointerDown(PointerEventData eventData)
     {
-
     }
     public void OnDrop(PointerEventData eventData)
     {
