--- Assets/Scripts/GameExitController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 010c2aa75dc48ea4198794762b45b1a4
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
