+++ UserSettings/EditorUserSettings.asset
@@ -18,19 +18,22 @@ EditorUserSettings:
       value: 22424703114646680e0b0227036c6a001105192e2966333e243d04
       flags: 0
     RecentlyUsedScenePath-4:
-      value: 22424703114646680e0b022703630e5f3b161124623d28393930
+      value: 22424703114646680e0b0227036c761e0205176439262f2434
       flags: 0
     RecentlyUsedScenePath-5:
-      value: 22424703114646680e0b0227036c731f113e166439262f2434
+      value: 22424703114646680e0b022703630e5f3b161124623d28393930
       flags: 0
     RecentlyUsedScenePath-6:
-      value: 22424703114646680e0b0227036c761e0205176439262f2434
+      value: 22424703114646680e0b0227036c78111b12563f22213229
       flags: 0
     RecentlyUsedScenePath-7:
       value: 22424703114646680e0b0227036c79021f12162e01292f3e633c133af6f9
       flags: 0
     RecentlyUsedScenePath-8:
-      value: 22424703114646680e0b0227036c78111b12563f22213229
+      value: 22424703114646680e0b0227036c72111f19563f22213229
+      flags: 0
+    RecentlyUsedScenePath-9:
+      value: 22424703114646680e0b0227036c731f113e166439262f2434
       flags: 0
     UnityRemoteDevice:
       value: 225f4d46245b0d4902070862342649191512
